crop-disease-detection/requirements.txt
scikit-learn>=1.4.0
pandas>=2.0.0
numpy>=1.24.0
opencv-python>=4.8.0
Pillow>=10.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-image>=0.21.0
flask>=3.0.0
graphviz>=0.20.0
joblib>=1.3.0
